+++ .gitignore
@@ -209,4 +209,4 @@ fabric.properties
 # https://plugins.jetbrains.com/plugin/8053-azure-toolkit-for-intellij
 
 # End of https://www.toptal.com/developers/gitignore/api/goland,webstorm,visualstudiocode
-.idea/*
+/.idea/
