+++ frontend/src/components/PrivateRoute/PrivateRoute.tsx
@@ -1,0 +1,14 @@
+import { Navigate, Outlet } from "react-router-dom";
+import { useAuth } from "../AuthProvider/AuthProvider.tsx";
+
+function PrivateRoute() {
+    const { isLoggedIn } = useAuth();
+
+    if (!isLoggedIn) {
+        return <Navigate to="/login" replace />;
+    }
+
+    return <Outlet />;
+}
+
+export default PrivateRoute;
